deply-app/requirements.txt
airtable-python-wrapper==0.15.3
asgiref==3.4.1
autopep8==1.6.0
boto3==1.23.3
botocore==1.26.3
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.11
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.11.0
django-eventlog==1.1
django-formtools==2.3
django-magiclink==1.2.0
django-multiselectfield==0.1.12
django-otp==1.1.3
django-phonenumber-field==6.3.0
django-prometheus==2.2.0
django-ses==3.0.1
django-two-factor-auth==1.14.0
gunicorn==20.1.0
idna==3.3
install==1.3.5
jmespath==1.0.0
packaging==20.9
phonenumbers==8.12.50
prometheus-client==0.14.1
pycodestyle==2.8.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
qrcode==7.3.1
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
urllib3==1.26.9
whitenoise==5.3.0
